chore: remove old files and move hook to separate file
